requirements.txt
python-telegram-bot==13.7
redis==4.5.1
python-dotenv==0.19.0 